restructure cli files
